Practica4

+++ practica3/finmundo.c
@@ -4,7 +4,7 @@
 
 int main()
 {
-	time_t longt = LONG_MAX; //Basta con esta linea	
+	time_t longt = INT_MAX; //Basta con esta linea	
 	printf("\nEl fin del mundo es: %s\n",ctime(&longt)); //Y esta 
 }
